Crate bp_xcm_bridge_hub

Primitives of the xcm-bridge-hub pallet.

Structs

Bridge

Bridge metadata.

BridgeId

Bridge identifier - used only for communicating with sibling/parent chains in the same consensus.

BridgeLocations

Locations of bridge endpoints at both sides of the bridge.

Enums

BridgeLocationsError

Errors that may happen when we check bridge locations.

BridgeState

Bridge state.

XcmBridgeHubCall

A minimized version of pallet_xcm_bridge_hub::Call that can be used without a runtime.

Traits

LocalXcmChannelManager

Local XCM channel manager.

Type Aliases

XcmAsPlainPayload

Encoded XCM blob. We expect the bridge messages pallet to use this blob type for both inbound and outbound payloads.